+++ packages/data-context/test/unit/sources/GitDataSource.spec.ts
@@ -43,7 +43,8 @@ describe('GitDataSource', () => {
     sinon.restore()
   })
 
-  it(`gets correct status for files on ${os.platform()}`, async () => {
+  // TODO: fix flaky test https://github.com/cypress-io/cypress/issues/23317
+  it.skip(`gets correct status for files on ${os.platform()}`, async () => {
     const onBranchChange = sinon.stub()
     const onGitInfoChange = sinon.stub()
     const onError = sinon.stub()