+++ system-tests/projects/experimentalSingleTabRunMode/src/2_foo.cy.js
@@ -1,0 +1,5 @@
+describe('component', () => {
+  it('passes', () => {
+    expect('foo').to.eq('foo')
+  })
+})